// src/types/profile-cafe.d.ts
export type ProfileCafeState = {
  status?: string;
  errors?: {
    id?: string[];
    tagline?: string[];
    description?: string[];
    banner_url?: string[];
    open_time?: string[];
    close_time?: string[];
    wifi_speed?: string[];
    total_sockets?: string[];
    _form?: string[];
  };
};